# api/openapi.yaml
# yaml-language-server: $schema=https://raw.githubusercontent.com/OAI/OpenAPI-Specification/main/schemas/v3.0/schema.json
openapi: "3.0.0"
info:
  version: "1.0.0"
  title: "URL Shortener API"
  description: "A REST API for shortening URLs, managing them, and handling redirects."
servers:
  - url: http://localhost:8080
paths:
  /urls:
    post:
      summary: "Create a shortened URL"
      operationId: "createShortUrl"
      tags:
        - "URL Management"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: "object"
              properties:
                originalUrl:
                  type: "string"
                  format: "uri"
                expiry:
                  type: "string"
                  format: "date-time"
              required:
                - "originalUrl"
      responses:
        '201':
          description: "Shortened URL created"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  shortUrl:
                    type: "string"
        '400':
          description: "Invalid request payload"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '409':
          description: "Conflict - URL already shortened"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '429':
          description: "Too many requests - Rate limit exceeded"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '500':
          description: "Internal server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /urls/{short-path}:
    get:
      summary: "Retrieve details of a shortened URL"
      operationId: "getShortUrlDetails"
      tags:
        - "URL Management"
      parameters:
        - name: "short-path"
          in: "path"
          required: true
          schema:
            type: "string"
      responses:
        '200':
          description: "URL details retrieved"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShortenedUrlDetails"
        '404':
          description: "URL not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '429':
          description: "Too many requests - Rate limit exceeded"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '500':
          description: "Internal server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      summary: "Update a shortened URL"
      operationId: "updateShortUrl"
      tags:
        - "URL Management"
      parameters:
        - name: "short-path"
          in: "path"
          required: true
          schema:
            type: "string"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: "object"
              properties:
                originalUrl:
                  type: "string"
                  format: "uri"
                expiry:
                  type: "string"
                  format: "date-time"
              required:
                - "originalUrl"
      responses:
        '200':
          description: "Shortened URL updated"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShortenedUrlDetails"
        '400':
          description: "Invalid request payload"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '404':
          description: "URL not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '409':
          description: "Conflict - URL update conflict"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '500':
          description: "Internal server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      summary: "Delete a shortened URL"
      operationId: "deleteShortUrl"
      tags:
        - "URL Management"
      parameters:
        - name: "short-path"
          in: "path"
          required: true
          schema:
            type: "string"
      responses:
        '204':
          description: "URL deleted successfully"
        '404':
          description: "URL not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '429':
          description: "Too many requests - Rate limit exceeded"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '500':
          description: "Internal server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /{short-path}:
    get:
      summary: "Redirect to the original URL"
      operationId: "redirectToOriginalUrl"
      tags:
        - "Redirect"
      parameters:
        - name: "short-path"
          in: "path"
          required: true
          schema:
            type: "string"
      responses:
        '302':
          description: "Redirecting to original URL"
          headers:
            Location:
              schema:
                type: "string"
              description: "URL to redirect to"
        '404':
          description: "URL not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '429':
          description: "Too many requests - Rate limit exceeded"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '500':
          description: "Internal server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /urls/{short-path}/stats:
    get:
      summary: "Get access statistics for a shortened URL"
      operationId: "getShortUrlStats"
      tags:
        - "Statistics"
      parameters:
        - name: "short-path"
          in: "path"
          required: true
          schema:
            type: "string"
      responses:
        '200':
          description: "Access statistics retrieved"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/URLStatistics"
        '404':
          description: "Shortened URL not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '429':
          description: "Too many requests - Rate limit exceeded"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '500':
          description: "Internal server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    ShortenedUrlDetails:
      type: "object"
      properties:
        short-path:
          type: "string"
        originalUrl:
          type: "string"
        shortUrl:
          type: "string"
        expiry:
          type: "string"
          format: "date-time"
    URLStatistics:
      type: "object"
      properties:
        last24Hours:
          type: "integer"
          description: "Number of accesses in the last 24 hours"
        pastWeek:
          type: "integer"
          description: "Number of accesses in the past week"
        allTime:
          type: "integer"
          description: "Total number of accesses"
    ErrorResponse:
      type: "object"
      properties:
        message:
          type: "string"