# generated/models.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.4.1 DO NOT EDIT.
package api

import (
	"time"
)

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	Message *string `json:"message,omitempty"`
}

// ShortenedUrlDetails defines model for ShortenedUrlDetails.
type ShortenedUrlDetails struct {
	Expiry      *time.Time `json:"expiry,omitempty"`
	OriginalUrl *string    `json:"originalUrl,omitempty"`
	ShortPath   *string    `json:"short-path,omitempty"`
	ShortUrl    *string    `json:"shortUrl,omitempty"`
}

// URLStatistics defines model for URLStatistics.
type URLStatistics struct {
	// AllTime Total number of accesses
	AllTime *int `json:"allTime,omitempty"`

	// Last24Hours Number of accesses in the last 24 hours
	Last24Hours *int `json:"last24Hours,omitempty"`

	// PastWeek Number of accesses in the past week
	PastWeek *int `json:"pastWeek,omitempty"`
}

// CreateShortUrlJSONBody defines parameters for CreateShortUrl.
type CreateShortUrlJSONBody struct {
	Expiry      *time.Time `json:"expiry,omitempty"`
	OriginalUrl string     `json:"originalUrl"`
}

// UpdateShortUrlJSONBody defines parameters for UpdateShortUrl.
type UpdateShortUrlJSONBody struct {
	Expiry      *time.Time `json:"expiry,omitempty"`
	OriginalUrl string     `json:"originalUrl"`
}

// CreateShortUrlJSONRequestBody defines body for CreateShortUrl for application/json ContentType.
type CreateShortUrlJSONRequestBody CreateShortUrlJSONBody

// UpdateShortUrlJSONRequestBody defines body for UpdateShortUrl for application/json ContentType.
type UpdateShortUrlJSONRequestBody UpdateShortUrlJSONBody